Move trajectory_interface to a separate directory.
General-purpose trajectory_interface functionality should live in a separate
place (ie. ROS package) than the joint_trajectory_controller. This is a first
step towards this separation.
