+++ joint_trajectory_controller/test/quintic_spline_segment_test.cpp
@@ -31,7 +31,7 @@
 #include <stdexcept>
 #include <gtest/gtest.h>
 #include <ros/console.h>
-#include <joint_trajectory_controller/quintic_spline_segment.h>
+#include <trajectory_interface/quintic_spline_segment.h>
 
 using namespace trajectory_interface;
